New menus, new front page, thorough testing, & general shininess
This one was thoroughly tested at all browser sizes, Firefox and
Webkit, anonymous and logged-in. Among other things, this commit adds
the new three-image front page jumbotron, fixes the issue with parts of
the menu blinking/disappearing when resizing horizontally
smaller-larger-smaller; placement of that rascally .field-quote on all
concerned pages, and placement of moderation/favorite icons in correct
place regardless of their toggle status. This is a big step forward and
I'm okay with not actively working on this theme (except bug fixes)
again until May if needed.